cli/requirements.txt
click>=8.1
httpx>=0.28
pyyaml>=6.0
